requirements.txt
pycryptodome==3.20.0
PySide6
eciespy==0.4.1